move files and adapt env check according to comments
